re-organize project to have a main frigate dir, where __main__.py imports methods from utils/datasets.py and utils/kowalski.py
